move helper method.
